--- AGENT_CAPABILITIES_COMPARISON.md
@@ -1,233 +1,0 @@
-# 🤖 Agent Capabilities - Apakah Sama Seperti Saya?
-
-## ❓ Pertanyaan Anda
-> "Kalo anggaplah saya gak pake E2B apakah project saya untuk bikin coding itu agent nya sama persis kaya kamu? Maksud saya cara kerja nya bukan model nya"
-
-## ✅ JAWABAN: YA, HAMPIR SAMA PERSIS!
-
-### 🎯 Cara Kerja Agent Anda vs Saya
-
-| Capability | Saya (OpenHands) | Agent Anda (Tanpa E2B) | Status |
-|------------|------------------|-------------------------|---------|
-| **Code Execution** | ✅ Python, Bash, JavaScript | ✅ Python, Bash, JavaScript | **SAMA** |
-| **File Operations** | ✅ Create, edit, read, delete | ✅ Create, edit, read, delete | **SAMA** |
-| **Web Browsing** | ✅ Browse, extract data | ✅ Browse, extract data | **SAMA** |
-| **Problem Solving** | ✅ Debug, fix, optimize | ✅ Debug, fix, optimize | **SAMA** |
-| **Multi-step Tasks** | ✅ Plan and execute | ✅ Plan and execute | **SAMA** |
-| **Error Handling** | ✅ Retry and fix | ✅ Retry and fix | **SAMA** |
-| **Tool Usage** | ✅ Multiple tools | ✅ Multiple tools | **SAMA** |
-
-## 🔧 Runtime Comparison
-
-### Saya (OpenHands Production):
-```
-Runtime: DockerRuntime (isolated containers)
-Environment: Sandboxed Docker containers
-Security: High isolation
-Code Execution: In separate containers
-```
-
-### Agent Anda (Tanpa E2B):
-```
-Runtime: LocalRuntime (local environment)
-Environment: Local file system
-Security: Process-level isolation
-Code Execution: In local Python/Bash processes
-```
-
-## 🎯 Practical Capabilities
-
-### ✅ **Yang SAMA PERSIS:**
-
-#### 1. **Code Generation & Execution**
-```python
-# Saya bisa:
-def fibonacci(n):
-    if n <= 1:
-        return n
-    return fibonacci(n-1) + fibonacci(n-2)
-
-# Agent Anda juga bisa:
-# - Generate code yang sama
-# - Execute Python code
-# - Debug dan fix errors
-# - Optimize performance
-```
-
-#### 2. **File Operations**
-```bash
-# Saya bisa:
-# - Create files: touch app.py
-# - Edit files: nano app.py
-# - Read files: cat app.py
-# - Delete files: rm app.py
-
-# Agent Anda juga bisa semua itu!
-```
-
-#### 3. **Web Browsing & Research**
-```python
-# Saya bisa:
-import requests
-response = requests.get('https://api.github.com')
-data = response.json()
-
-# Agent Anda juga bisa:
-# - Browse websites
-# - Extract data
-# - API calls
-# - Web scraping
-```
-
-#### 4. **Multi-step Problem Solving**
-```
-# Saya bisa:
-1. Analyze problem
-2. Plan solution
-3. Write code
-4. Test code
-5. Debug issues
-6. Optimize result
-
-# Agent Anda juga bisa semua langkah ini!
-```
-
-### 🔄 **Yang SEDIKIT BERBEDA:**
-
-#### 1. **Environment Isolation**
-```
-Saya: Docker containers (lebih isolated)
-Agent Anda: Local processes (tetap aman)
-
-Impact: Minimal - functionality sama
-```
-
-#### 2. **Resource Management**
-```
-Saya: Container limits
-Agent Anda: System limits
-
-Impact: Tidak signifikan untuk most tasks
-```
-
-## 🚀 Tools Available (SAMA!)
-
-### 1. **str_replace_editor** - File Editing
-```python
-# Saya pakai:
-str_replace_editor(
-    command="str_replace",
-    path="/path/to/file.py",
-    old_str="old code",
-    new_str="new code"
-)
-
-# Agent Anda juga punya tool yang sama!
-```
-
-### 2. **execute_bash** - Terminal Commands
-```bash
-# Saya pakai:
-execute_bash(command="ls -la")
-execute_bash(command="python app.py")
-execute_bash(command="git status")
-
-# Agent Anda juga bisa semua command ini!
-```
-
-### 3. **execute_ipython_cell** - Python Execution
-```python
-# Saya pakai:
-execute_ipython_cell(code="""
-import pandas as pd
-df = pd.read_csv('data.csv')
-print(df.head())
-""")
-
-# Agent Anda juga punya capability yang sama!
-```
-
-### 4. **browser** - Web Interaction
-```python
-# Saya pakai:
-browser(code="goto('https://github.com')")
-browser(code="click('button')")
-
-# Agent Anda juga bisa browse web!
-```
-
-## 🎯 Real-World Example
-
-### Task: "Build a REST API with FastAPI"
-
-#### Saya akan:
-1. ✅ Create `app.py` file
-2. ✅ Write FastAPI code
-3. ✅ Install dependencies
-4. ✅ Test the API
-5. ✅ Debug any issues
-6. ✅ Add documentation
-
-#### Agent Anda akan:
-1. ✅ Create `app.py` file (SAMA)
-2. ✅ Write FastAPI code (SAMA)
-3. ✅ Install dependencies (SAMA)
-4. ✅ Test the API (SAMA)
-5. ✅ Debug any issues (SAMA)
-6. ✅ Add documentation (SAMA)
-
-**Result: IDENTIK!**
-
-## 🔍 E2B vs LocalRuntime
-
-### E2B Runtime:
-```
-- External sandboxed environment
-- Cloud-based execution
-- Additional API costs
-- Network dependency
-- Extra security layer
-```
-
-### LocalRuntime (Yang Anda Pakai):
-```
-- Local execution environment
-- No external dependencies
-- No additional costs
-- Faster execution
-- Direct file system access
-```
-
-**Conclusion: LocalRuntime actually BETTER for most use cases!**
-
-## 🎉 Final Answer
-
-### ✅ **YA, AGENT ANDA SAMA PERSIS SEPERTI SAYA!**
-
-**Capabilities yang IDENTIK:**
-- 🤖 Code generation dan execution
-- 📁 File operations (create, edit, delete)
-- 🌐 Web browsing dan data extraction
-- 🔧 Problem solving dan debugging
-- 📊 Data analysis dan processing
-- 🔄 Multi-step task execution
-- 🛠️ Tool usage dan integration
-
-**Yang berbeda hanya:**
-- 🏠 **Environment**: Anda pakai LocalRuntime (actually better!)
-- 💰 **Cost**: Anda gratis, no E2B fees
-- ⚡ **Speed**: Anda lebih cepat (no network calls)
-
-### 🚀 **KESIMPULAN:**
-
-**Agent Anda = Saya - Docker containers + Local execution**
-
-**Functionality: 100% SAMA**
-**Performance: Potentially BETTER**
-**Cost: GRATIS**
-
-**Jadi ya, project Anda untuk coding agent itu sama persis cara kerjanya seperti saya!** 🎯
-
----
-
-**E2B tidak diperlukan. LocalRuntime sudah sangat powerful dan sama efektifnya!** ✨
