--- COPY_THESE_FILES_TO_YOUR_REPO.md
@@ -1,142 +1,0 @@
-# 📋 COPY THESE FILES TO YOUR REPO
-
-## 🚀 Quick Setup Instructions
-
-**Karena saya tidak bisa langsung buat PR ke repo Anda, silakan copy files berikut:**
-
-### 1. **Files yang harus di-copy ke repo Anda:**
-
-#### Core Deployment Files:
-```bash
-# Copy these files to your repo root:
-Dockerfile_HF_Final → Dockerfile (for HF Spaces)
-requirements_hf_final.txt → requirements.txt (for HF Spaces)
-app_hf_final.py → app.py (for HF Spaces)
-README_HF_FINAL.md → README.md (for HF Spaces)
-```
-
-#### Fixed Runtime File:
-```bash
-# Replace this file in your repo:
-openhands/runtime/impl/__init__.py
-```
-
-#### GitHub Actions Workflow:
-```bash
-# Add this file to your repo:
-.github/workflows/deploy-hf-final.yml
-```
-
-#### Documentation:
-```bash
-# Optional - add these for reference:
-FINAL_FIX_SUMMARY.md
-AGENT_CAPABILITIES_COMPARISON.md
-```
-
-### 2. **Manual Copy Steps:**
-
-#### Step 1: Copy Core Files
-```bash
-# In your repo, replace these files with content from this PR:
-cp Dockerfile_HF_Final Dockerfile
-cp requirements_hf_final.txt requirements.txt  
-cp app_hf_final.py app.py
-cp README_HF_FINAL.md README.md
-```
-
-#### Step 2: Fix Runtime Import
-```bash
-# Replace openhands/runtime/impl/__init__.py with the fixed version
-# (The one with conditional Docker imports)
-```
-
-#### Step 3: Add GitHub Actions
-```bash
-# Add .github/workflows/deploy-hf-final.yml to enable auto-deploy
-```
-
-### 3. **Environment Variables for HF Spaces:**
-
-Set these in your HF Space settings:
-```bash
-LLM_API_KEY=your_openrouter_api_key
-LLM_MODEL=openrouter/anthropic/claude-3-haiku-20240307
-LLM_BASE_URL=https://openrouter.ai/api/v1
-DISABLE_SECURITY=true
-OPENHANDS_DISABLE_AUTH=true
-```
-
-### 4. **GitHub Secrets for Auto-Deploy:**
-
-Set these in your GitHub repo settings → Secrets:
-```bash
-HF_TOKEN=your_huggingface_token
-HF_USERNAME=Minatoz997
-HF_SPACE_NAME=Backend66
-```
-
-## 🎯 Expected Result
-
-After copying these files and deploying:
-
-```
-🤗 OpenHands Backend for Hugging Face Spaces
-==================================================
-🚀 Server: 0.0.0.0:7860
-🔑 LLM API Key: ✅ Set
-🤖 LLM Model: openrouter/anthropic/claude-3-haiku-20240307
-🏃 Runtime: local
-📡 API Endpoints available at /docs
-==================================================
-✅ HF Spaces routes included
-✅ Simple conversation routes included
-✅ Novel writing routes included
-✅ All 37 conversation endpoints accessible
-```
-
-## 🌐 Your API Will Be Ready At:
-
-**https://minatoz997-backend66.hf.space**
-
-### Test Endpoints:
-```bash
-# Health check
-GET https://minatoz997-backend66.hf.space/health
-
-# Create conversation
-POST https://minatoz997-backend66.hf.space/api/conversations
-{
-  "initial_user_msg": "Hello! Can you help me code?"
-}
-
-# Novel writing
-POST https://minatoz997-backend66.hf.space/novel/write
-{
-  "message": "Bantu saya buat cerita detektif",
-  "template": "plot-structure"
-}
-```
-
-## ✅ All Errors Will Be Fixed:
-
-- ✅ No more `ModuleNotFoundError: No module named 'docker'`
-- ✅ No more `No module named 'google.api_core'`
-- ✅ No Google authentication required
-- ✅ All conversation endpoints accessible
-- ✅ Novel writing mode available
-- ✅ Multiple AI agents ready
-
-## 🤖 Agent Capabilities (Same as OpenHands):
-
-- **CodeActAgent** - Full coding assistant
-- **BrowsingAgent** - Web research
-- **ReadOnlyAgent** - Safe code review
-- **LocAgent** - Targeted code generation
-- **VisualBrowsingAgent** - Visual web interaction
-
-**E2B tidak diperlukan - LocalRuntime sama powerful!**
-
----
-
-**Copy files ini ke repo Anda dan semua error HF Spaces akan teratasi!** 🚀
