--- DEPLOY_HF_FINAL.md
@@ -1,185 +1,0 @@
-# 🤗 Hugging Face Spaces Deployment - Final Fixed Version
-
-## 🚨 Problem Solved
-
-This deployment fixes the critical import errors:
-- ❌ `ModuleNotFoundError: No module named 'docker'`
-- ❌ `No module named 'google.api_core'`
-- ✅ **All fixed!** Ready for HF Spaces deployment
-
-## 📋 Quick Deploy Steps
-
-### 1. Create New Space
-- Go to [huggingface.co/new-space](https://huggingface.co/new-space)
-- **Space name**: `openhands-backend` (or any name you prefer)
-- **SDK**: Docker
-- **Visibility**: Public
-- **Hardware**: CPU basic (free tier)
-
-### 2. Upload Fixed Files
-
-Upload these **FIXED** files to your Space:
-
-**Required files:**
-```
-Dockerfile_HF_Final → rename to → Dockerfile
-app_hf_final.py
-requirements_hf_fixed.txt → rename to → requirements.txt
-README.md
-openhands/ (entire folder)
-```
-
-**File mapping:**
-- `Dockerfile_HF_Final` → `Dockerfile`
-- `requirements_hf_fixed.txt` → `requirements.txt`
-- `app_hf_final.py` → keep as is (or rename to `app.py`)
-
-### 3. Set Environment Variables
-
-In your Space Settings → Environment Variables:
-
-```bash
-# Required for LLM functionality
-LLM_API_KEY=your_openrouter_api_key
-
-# Optional (already set as defaults)
-LLM_MODEL=openrouter/anthropic/claude-3-haiku-20240307
-LLM_BASE_URL=https://openrouter.ai/api/v1
-SESSION_API_KEY=random_string_123
-DEBUG=false
-```
-
-### 4. Deploy!
-- HF will automatically build and deploy
-- Your API will be available at: `https://your-username-space-name.hf.space`
-
-## 🔧 What Was Fixed
-
-### Import Issues Resolved:
-1. **Docker dependency**: Made conditional import in `openhands/runtime/impl/__init__.py`
-2. **Google Cloud**: Already had conditional imports, but warnings removed
-3. **Requirements**: Created `requirements_hf_fixed.txt` without problematic dependencies
-4. **App entry point**: Created `app_hf_final.py` with better error handling
-
-### Key Changes:
-
-#### 1. Fixed Runtime Imports (`openhands/runtime/impl/__init__.py`)
-```python
-# Before (caused error):
-from openhands.runtime.impl.docker.docker_runtime import DockerRuntime
-
-# After (conditional import):
-try:
-    from openhands.runtime.impl.docker.docker_runtime import DockerRuntime
-    DOCKER_AVAILABLE = True
-except ImportError:
-    class DockerRuntime:
-        def __init__(self, *args, **kwargs):
-            raise ImportError("DockerRuntime requires docker package. Use LocalRuntime instead.")
-    DOCKER_AVAILABLE = False
-```
-
-#### 2. Cleaned Requirements (`requirements_hf_fixed.txt`)
-**Removed problematic dependencies:**
-- `docker` - Not needed in HF Spaces containers
-- `google-generativeai` - Causes google.cloud conflicts
-- `redis` - External service
-- `e2b` - External service
-- `libtmux` - Terminal multiplexer issues
-- And many others...
-
-#### 3. Enhanced App Entry Point (`app_hf_final.py`)
-- Better dependency checking
-- Clearer error messages
-- Proper environment setup
-- Startup diagnostics
-
-## 🌐 API Endpoints
-
-Once deployed, test these endpoints:
-
-```bash
-# Health check
-GET https://your-space.hf.space/health
-
-# Get config
-GET https://your-space.hf.space/api/options/config
-
-# Create conversation
-POST https://your-space.hf.space/api/conversations
-Content-Type: application/json
-{
-  "initial_user_msg": "Hello! Can you help me with coding?"
-}
-```
-
-## 🎯 Frontend Integration
-
-Your deployed backend will work with frontends on:
-- **Vercel** (*.vercel.app)
-- **Netlify** (*.netlify.app)
-- **GitHub Pages** (*.github.io)
-- **Local development** (localhost)
-
-CORS is pre-configured to allow all origins.
-
-## 🚀 Example Usage
-
-```javascript
-// Create a new conversation
-const response = await fetch('https://your-space.hf.space/api/conversations', {
-  method: 'POST',
-  headers: { 'Content-Type': 'application/json' },
-  body: JSON.stringify({
-    initial_user_msg: 'Write a Python function to calculate fibonacci numbers'
-  })
-});
-
-const conversation = await response.json();
-console.log(conversation);
-```
-
-## 💡 Troubleshooting
-
-### If you still get import errors:
-1. Make sure you uploaded the **FIXED** files:
-   - `Dockerfile_HF_Final` as `Dockerfile`
-   - `requirements_hf_fixed.txt` as `requirements.txt`
-   - `app_hf_final.py`
-
-2. Check HF Spaces logs for specific errors
-
-3. Verify environment variables are set correctly
-
-### Common Issues:
-- **"No module named 'docker'"** → Use fixed files
-- **"Google Cloud dependencies not available"** → This is normal, just a warning
-- **"LLM_API_KEY not set"** → Add your OpenRouter API key to environment variables
-
-## 🎉 Success Indicators
-
-When deployment works, you'll see:
-```
-🤗 OpenHands Backend for Hugging Face Spaces
-==================================================
-🚀 Server: 0.0.0.0:7860
-🔑 LLM API Key: ✅ Set
-🤖 LLM Model: openrouter/anthropic/claude-3-haiku-20240307
-🏃 Runtime: local
-📡 API Endpoints available at /docs
-==================================================
-```
-
-## 📚 Next Steps
-
-1. **Deploy Backend**: Follow steps above
-2. **Get API Key**: Sign up at [OpenRouter](https://openrouter.ai)
-3. **Create Frontend**: Deploy UI on Vercel/Netlify
-4. **Connect**: Point frontend to your HF Space URL
-5. **Test**: Create conversations and chat with AI!
-
-Happy deploying! 🚀
-
----
-
-**Note**: This version is specifically optimized for Hugging Face Spaces and removes all problematic dependencies that cause import errors.
