--- Dockerfile_HF_Ultra_Minimal
@@ -1,46 +1,0 @@
-FROM python:3.11-slim
-
-WORKDIR /app
-
-# Install system dependencies
-RUN apt-get update && apt-get install -y \
-    git \
-    curl \
-    && rm -rf /var/lib/apt/lists/*
-
-# Clone repository
-RUN git clone https://github.com/Minatoz997/OpenHands-Backend.git .
-
-# Install Python dependencies (minimal set)
-RUN pip install --no-cache-dir \
-    fastapi==0.104.1 \
-    uvicorn[standard]==0.24.0 \
-    litellm==1.44.22 \
-    httpx==0.25.2 \
-    pydantic==2.5.0 \
-    python-multipart==0.0.6 \
-    python-dotenv==1.0.0 \
-    aiohttp==3.9.1 \
-    termcolor==2.4.0 \
-    toml==0.10.2 \
-    numpy==1.26.2 \
-    jinja2==3.1.3 \
-    tenacity==8.5.0 \
-    pathspec==0.12.1 \
-    pyjwt==2.9.0 \
-    binaryornot==0.4.4 \
-    psutil==5.9.6 \
-    requests==2.31.0 \
-    urllib3==2.1.0
-
-# Expose port
-EXPOSE 7860
-
-# Set environment variables
-ENV PYTHONPATH=/app
-ENV HF_SPACES=1
-ENV ENVIRONMENT=production
-ENV RUNTIME=cli
-
-# Start application
-CMD ["python", "app_hf_final.py"]
