--- Dockerfile_ultra_simple
@@ -1,22 +1,0 @@
-# Ultra Simple Dockerfile for HF Spaces
-FROM python:3.11-slim
-
-WORKDIR /app
-
-# Copy requirements and install
-COPY requirements_ultra_simple.txt requirements.txt
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Copy application
-COPY app_ultra_simple.py app.py
-
-# Expose port
-EXPOSE 7860
-
-# Set environment variables
-ENV HOST=0.0.0.0
-ENV PORT=7860
-ENV PYTHONUNBUFFERED=1
-
-# Run application
-CMD ["python", "app.py"]
