--- FIXES_DOCUMENTATION.md
@@ -1,133 +1,0 @@
-# OpenHands Backend - Hugging Face Spaces Fixes
-
-## Masalah yang Diperbaiki
-
-### 1. Error "Not Found" pada Root URL
-**Masalah**: Ketika mengakses root URL (`/`), server mengembalikan 404 Not Found.
-
-**Solusi**: 
-- Menambahkan root endpoint (`/`) yang memberikan informasi API
-- Endpoint ini menampilkan daftar endpoint yang tersedia dan informasi server
-
-### 2. Missing CORS Configuration
-**Masalah**: Request dari frontend mungkin diblokir karena CORS tidak dikonfigurasi dengan benar.
-
-**Solusi**:
-- Menambahkan `CORSMiddleware` dengan konfigurasi yang tepat
-- Mengizinkan semua origins untuk deployment publik di HF Spaces
-
-### 3. Poor Error Handling
-**Masalah**: Error 404 dan error lainnya tidak memberikan informasi yang berguna.
-
-**Solusi**:
-- Menambahkan global exception handlers untuk 404, validation errors, dan general exceptions
-- Error response sekarang memberikan informasi yang lebih berguna
-
-### 4. Missing Endpoints
-**Masalah**: Beberapa endpoint yang diminta oleh HF Spaces UI tidak tersedia.
-
-**Solusi**:
-- Menambahkan endpoint `/api/hf/logs-container` dan `/api/hf/logs`
-- Endpoint ini mengembalikan response yang sesuai untuk request dari HF Spaces
-
-### 5. Better Logging and Debugging
-**Masalah**: Sulit untuk debug masalah karena logging yang kurang informatif.
-
-**Solusi**:
-- Menambahkan logging yang lebih baik di `app_hf.py`
-- Menambahkan try-catch untuk menangani error startup
-- Menampilkan informasi debug yang lebih lengkap
-
-## Perubahan File
-
-### 1. `/openhands/server/app.py`
-- ✅ Menambahkan import untuk CORS dan error handling
-- ✅ Menambahkan CORSMiddleware
-- ✅ Menambahkan root endpoint (`/`)
-- ✅ Menambahkan global exception handlers
-
-### 2. `/app_hf.py`
-- ✅ Menambahkan logging yang lebih baik
-- ✅ Menambahkan error handling untuk startup
-- ✅ Menambahkan informasi debug yang lebih lengkap
-
-### 3. `/openhands/server/routes/hf_spaces.py`
-- ✅ Menambahkan endpoint `/logs-container` dan `/logs`
-- ✅ Menambahkan import untuk HTTPException dan JSONResponse
-
-### 4. `/test_fixes.py` (Baru)
-- ✅ Script untuk testing endpoint setelah perbaikan
-
-## Endpoint yang Tersedia Setelah Perbaikan
-
-### Root Endpoints
-- `GET /` - Informasi API dan daftar endpoint
-- `GET /health` - Health check
-- `GET /docs` - API documentation (Swagger UI)
-- `GET /openapi.json` - OpenAPI specification
-
-### API Endpoints
-- `GET /api/options/config` - Konfigurasi server
-- `GET /api/options/models` - Daftar model LLM yang tersedia
-- `GET /api/options/agents` - Daftar agent yang tersedia
-- `POST /api/conversations` - Membuat conversation baru
-
-### HF Spaces Specific
-- `GET /api/hf/status` - Status deployment HF Spaces
-- `GET /api/hf/ready` - Health check untuk HF Spaces
-- `GET /api/hf/environment` - Informasi environment variables
-- `GET /api/hf/debug` - Debug informasi sistem dan environment
-- `GET /api/hf/logs-container` - Endpoint untuk logs (placeholder)
-- `GET /api/hf/logs` - Endpoint untuk logs (placeholder)
-- `POST /api/hf/test-conversation` - Test endpoint untuk membuat conversation sederhana
-
-## Testing
-
-Untuk menguji perbaikan:
-
-```bash
-# Install dependencies
-pip install -r requirements.txt
-
-# Run test script
-python test_fixes.py
-
-# Atau jalankan server manual
-python app_hf.py
-```
-
-## Deployment ke Hugging Face Spaces
-
-1. Pastikan file `space_config.yml` sudah benar
-2. Pastikan `Dockerfile_HF` sudah benar
-3. Push perubahan ke repository
-4. HF Spaces akan otomatis rebuild dan deploy
-
-## Environment Variables untuk HF Spaces
-
-Variabel environment yang direkomendasikan untuk HF Spaces:
-
-```
-OPENHANDS_RUNTIME=local
-CORS_ALLOWED_ORIGINS=*
-DISABLE_SECURITY=true
-OPENHANDS_DISABLE_AUTH=true
-SETTINGS_STORE_TYPE=memory
-SECRETS_STORE_TYPE=memory
-SKIP_SETTINGS_MODAL=true
-```
-
-## Troubleshooting
-
-### Jika masih ada error 404:
-1. Periksa apakah server sudah start dengan benar
-2. Cek logs untuk error startup
-3. Pastikan semua dependencies terinstall
-
-### Jika ada CORS error:
-1. Pastikan `CORS_ALLOWED_ORIGINS=*` di environment
-2. Periksa apakah CORSMiddleware sudah ditambahkan
-
-### Jika ada import error:
-1. Pastikan PYTHONPATH sudah benar
-2. Periksa apakah semua dependencies ada di requirements.txt
