--- MANUAL_CONTRIBUTION_GUIDE.md
@@ -1,191 +1,0 @@
-# 📋 MANUAL CONTRIBUTION GUIDE - DOCKER FIXES
-
-## 🎯 **FILES TO COPY TO YOUR MAIN REPOSITORY**
-
-### **1. Modified Files (Replace existing content):**
-
-#### **A. `openhands/runtime/__init__.py`**
-- **Action:** Replace entire file content
-- **Location:** `/workspace/OpenHands-Backend/OpenHands-Backend/openhands/runtime/__init__.py`
-- **Changes:** Added fallback system for all runtime imports
-
-#### **B. `openhands/agenthub/browsing_agent/__init__.py`**
-- **Action:** Replace entire file content  
-- **Location:** `/workspace/OpenHands-Backend/OpenHands-Backend/openhands/agenthub/browsing_agent/__init__.py`
-- **Changes:** Added fallback for browsergym dependencies
-
-#### **C. `openhands/agenthub/visualbrowsing_agent/__init__.py`**
-- **Action:** Replace entire file content
-- **Location:** `/workspace/OpenHands-Backend/OpenHands-Backend/openhands/agenthub/visualbrowsing_agent/__init__.py`
-- **Changes:** Added fallback for browsergym dependencies
-
-#### **D. `COMPLETE_AI_ASSISTANT_PROMPT.md`**
-- **Action:** Replace entire file content
-- **Location:** `/workspace/OpenHands-Backend/OpenHands-Backend/COMPLETE_AI_ASSISTANT_PROMPT.md`
-- **Changes:** Updated with latest fix status
-
-### **2. New Files (Create new):**
-
-#### **A. `requirements_hf_minimal.txt`**
-- **Action:** Create new file
-- **Location:** Root directory
-- **Purpose:** Ultra-minimal dependencies for HF Spaces
-
-#### **B. `Dockerfile_HF_Ultra_Minimal`**
-- **Action:** Create new file
-- **Location:** Root directory  
-- **Purpose:** Optimized Dockerfile for HF Spaces
-
-#### **C. `DOCKER_IMPORT_FIXES_SUMMARY.md`**
-- **Action:** Create new file
-- **Location:** Root directory
-- **Purpose:** Complete documentation of docker fixes
-
-#### **D. `MANUAL_CONTRIBUTION_GUIDE.md`**
-- **Action:** Create new file (this file)
-- **Location:** Root directory
-- **Purpose:** Guide for manual contribution
-
----
-
-## 🔧 **STEP-BY-STEP CONTRIBUTION PROCESS**
-
-### **Step 1: Copy Modified Files**
-
-1. **Copy `openhands/runtime/__init__.py`:**
-   ```bash
-   # Copy content from this file to your repository
-   # Location: openhands/runtime/__init__.py
-   ```
-
-2. **Copy `openhands/agenthub/browsing_agent/__init__.py`:**
-   ```bash
-   # Copy content from this file to your repository
-   # Location: openhands/agenthub/browsing_agent/__init__.py
-   ```
-
-3. **Copy `openhands/agenthub/visualbrowsing_agent/__init__.py`:**
-   ```bash
-   # Copy content from this file to your repository
-   # Location: openhands/agenthub/visualbrowsing_agent/__init__.py
-   ```
-
-4. **Copy `COMPLETE_AI_ASSISTANT_PROMPT.md`:**
-   ```bash
-   # Copy content from this file to your repository
-   # Location: COMPLETE_AI_ASSISTANT_PROMPT.md
-   ```
-
-### **Step 2: Add New Files**
-
-1. **Create `requirements_hf_minimal.txt`**
-2. **Create `Dockerfile_HF_Ultra_Minimal`**
-3. **Create `DOCKER_IMPORT_FIXES_SUMMARY.md`**
-4. **Create `MANUAL_CONTRIBUTION_GUIDE.md`**
-
-### **Step 3: Test Before Commit**
-
-```bash
-# Test imports
-python -c "from openhands.runtime import CLIRuntime; print('✅ Runtime OK')"
-python -c "import openhands.agenthub; print('✅ Agenthub OK')"
-python -c "from openhands.server.app import app; print('✅ Server OK')"
-
-# Test server startup
-python app_hf_final.py
-# Should start without docker errors
-```
-
-### **Step 4: Commit and Push**
-
-```bash
-git add .
-git commit -m "🔧 Fix HF Spaces Docker Import Errors - Ultra Minimal Version
-
-✅ FIXES IMPLEMENTED:
-- Fixed docker import errors with fallback runtime system
-- Added fallback for browsing_agent and visualbrowsing_agent
-- Created ultra-minimal requirements (requirements_hf_minimal.txt)
-- Updated runtime/__init__.py with HF Spaces compatible imports
-- All imports now use CLIRuntime as fallback for HF Spaces
-
-✅ TESTED AND WORKING:
-- Server starts successfully without docker errors
-- All core agents available (CodeActAgent, ReadOnlyAgent, etc.)
-- Novel writing functionality intact
-- API endpoints functional
-
-🎯 PERSONAL PROJECT READY:
-- No Google login required
-- OpenRouter integration ready
-- Perfect for personal use with girlfriend
-- Cost-effective solution
-
-📦 NEW FILES:
-- Dockerfile_HF_Ultra_Minimal: Optimized for HF Spaces
-- requirements_hf_minimal.txt: Only essential dependencies
-
-🚀 DEPLOYMENT READY FOR HF SPACES!"
-
-git push origin main
-```
-
----
-
-## 🧪 **VERIFICATION CHECKLIST**
-
-### **Before Deployment:**
-- [ ] All modified files copied correctly
-- [ ] All new files created
-- [ ] Import test passes
-- [ ] Server starts without errors
-- [ ] No docker import errors
-- [ ] All endpoints accessible
-
-### **After Deployment to HF Spaces:**
-- [ ] Use `Dockerfile_HF_Ultra_Minimal`
-- [ ] Set OpenRouter API key
-- [ ] Test `/health` endpoint
-- [ ] Test `/docs` endpoint
-- [ ] Test novel writing feature
-- [ ] Test coding assistance
-
----
-
-## 🎉 **EXPECTED RESULTS**
-
-### **Successful Startup Log:**
-```
-🔧 Setting up Hugging Face environment...
-✅ Environment configured for Hugging Face Spaces
-🔍 Checking dependencies...
-✅ FastAPI available
-✅ Uvicorn available
-✅ LiteLLM available
-📦 Importing OpenHands app...
-✅ HF Spaces routes included
-✅ Novel writing routes included
-🚀 Starting uvicorn server...
-INFO: Uvicorn running on http://0.0.0.0:7860
-```
-
-### **Working Features:**
-- ✅ AI coding assistance (CodeActAgent, ReadOnlyAgent, LocAgent)
-- ✅ Indonesian novel writing (7 templates)
-- ✅ File management (upload, edit, organize)
-- ✅ Cost-effective OpenRouter integration
-- ✅ Personal use optimization (no enterprise features)
-
----
-
-## 💕 **PERSONAL PROJECT SUCCESS**
-
-Your OpenHands backend is now ready for you and your girlfriend to use:
-
-1. **Coding Projects:** Get help with Python, JavaScript, debugging
-2. **Creative Writing:** Indonesian novel writing with AI assistance
-3. **File Management:** Organize and edit project files
-4. **Cost-Effective:** OpenRouter API instead of expensive alternatives
-5. **No Google Login:** Complete independence from Google services
-
-**Perfect for personal use! 🚀💕**
