--- requirements_hf.txt
@@ -1,83 +1,0 @@
-# Optimized requirements for Hugging Face Spaces
-# Core FastAPI and server dependencies
-fastapi>=0.104.0
-uvicorn[standard]>=0.24.0
-python-multipart
-python-dotenv
-
-# LLM and AI dependencies
-litellm>=1.60.0,!=1.64.4
-aiohttp>=3.9.0,!=3.11.13
-google-generativeai
-anthropic
-openai
-
-# Core utilities
-termcolor
-toml
-types-toml
-numpy
-json-repair
-html2text
-pexpect
-jinja2>=3.1.3
-tenacity>=8.5,<10.0
-pathspec>=0.12.1
-pyjwt>=2.9.0
-dirhash
-tornado
-rapidfuzz>=3.9.0
-whatthepatch>=1.0.6
-protobuf>=5.0.0,<6.0.0
-
-# Monitoring and telemetry
-opentelemetry-api>=1.33.1
-opentelemetry-exporter-otlp-proto-grpc>=1.33.1
-
-# GitHub integration
-pygithub>=2.5.0
-
-# Job processing
-joblib
-
-# WebSocket and SSE support
-python-socketio>=5.11.4
-sse-starlette>=2.1.3
-
-# System utilities
-psutil
-python-json-logger>=3.2.1
-prompt-toolkit>=3.0.50
-anyio==4.9.0
-
-# MCP support
-fastmcp>=2.5.2
-
-# Content processing
-python-frontmatter>=1.1.0
-PyPDF2
-python-pptx
-pylatexenc
-python-docx
-bashlex>=0.18
-
-# Cloud storage (optional)
-boto3
-
-# Remove problematic dependencies for HF Spaces:
-# - docker (not needed in containerized environment)
-# - pythonnet (Windows-specific)
-# - modal (external cloud service)
-# - daytona-sdk (not needed)
-# - poetry (not needed at runtime)
-# - qtconsole (GUI dependency)
-# - ipywidgets (Jupyter dependency)
-# - redis (external service)
-# - e2b (external service)
-# - google-cloud-aiplatform (heavy dependency)
-# - minio (external service)
-# - stripe (payment processing)
-# - libtmux (terminal multiplexer)
-# - zope-interface (specific version conflict)
-# - runloop-api-client (external service)
-# - openhands-aci (Azure specific)
