--- requirements_hf_fixed.txt
@@ -1,107 +1,0 @@
-# Optimized requirements for Hugging Face Spaces - Fixed Version
-# Removes problematic dependencies that cause import errors
-
-# ===== Core FastAPI and Server =====
-fastapi>=0.104.0
-uvicorn[standard]>=0.24.0
-python-multipart
-python-dotenv
-
-# ===== LLM and AI Dependencies =====
-litellm>=1.60.0,!=1.64.4,!=1.67.*
-aiohttp>=3.9.0,!=3.11.13
-anthropic
-openai
-
-# ===== Core Utilities =====
-termcolor
-toml
-types-toml
-numpy
-json-repair
-html2text
-pexpect
-jinja2>=3.1.3
-tenacity>=8.5,<10.0
-pathspec>=0.12.1
-pyjwt>=2.9.0
-dirhash
-tornado
-rapidfuzz>=3.9.0
-whatthepatch>=1.0.6
-protobuf>=5.0.0,<6.0.0
-
-# ===== Browser Automation (Required by browsing_agent) =====
-browsergym-core==0.13.3
-
-# ===== Monitoring and Telemetry =====
-opentelemetry-api>=1.33.1
-opentelemetry-exporter-otlp-proto-grpc>=1.33.1
-
-# ===== GitHub Integration =====
-pygithub>=2.5.0
-
-# ===== Job Processing =====
-joblib
-
-# ===== WebSocket and SSE =====
-python-socketio>=5.11.4
-sse-starlette>=2.1.3
-
-# ===== System Utilities =====
-psutil
-python-json-logger>=3.2.1
-prompt-toolkit>=3.0.50
-anyio==4.9.0
-
-# ===== MCP Support =====
-fastmcp>=2.5.2
-
-# ===== Content Processing =====
-python-frontmatter>=1.1.0
-PyPDF2
-python-pptx
-pylatexenc
-python-docx
-bashlex>=0.18
-Pillow
-
-# ===== File and Data Processing =====
-binaryornot
-deprecated
-
-# ===== Cloud Storage =====
-boto3
-botocore
-
-# ===== HTTP and Network =====
-requests
-urllib3
-charset-normalizer
-certifi
-idna
-
-# ===== EXCLUDED DEPENDENCIES (Cause import errors in HF Spaces) =====
-# docker - Causes "No module named 'docker'" error
-# google-generativeai - Causes google.cloud conflicts
-# google-api-python-client - Causes google.cloud storage conflicts
-# google-auth-httplib2 - Not needed without Google APIs  
-# google-auth-oauthlib - Not needed without Google APIs
-# google-cloud-aiplatform - Very heavy dependency, causes conflicts
-# redis - External service, not essential for basic functionality
-# e2b - External service, not essential for basic functionality
-# libtmux - Terminal multiplexer, causes issues in containers
-# zope-interface - Specific version conflicts
-# pythonnet - Windows-specific, causes build issues  
-# modal - External cloud service, heavy dependency
-# daytona-sdk - External service
-# poetry - Not needed at runtime
-# qtconsole - GUI dependency, not needed for API
-# ipywidgets - Jupyter dependency, not needed
-# minio - External service, not essential
-# stripe - Payment processing, not needed
-# runloop-api-client - External service
-# openhands-aci - Azure specific
-
-# NOTE: This configuration prioritizes stability and compatibility
-# over advanced features for HF Spaces deployment
