--- requirements_hf_minimal.txt
@@ -1,58 +1,0 @@
-# Minimal requirements for Hugging Face Spaces - Ultra Stable Version
-# Only essential dependencies to avoid import errors
-
-# ===== Core FastAPI and Server =====
-fastapi==0.104.1
-uvicorn[standard]==0.24.0
-python-multipart==0.0.6
-python-dotenv==1.0.0
-
-# ===== LLM and AI Dependencies =====
-litellm==1.44.22
-httpx==0.25.2
-aiohttp==3.9.1
-
-# ===== Core Utilities =====
-pydantic==2.5.0
-termcolor==2.4.0
-toml==0.10.2
-numpy==1.26.2
-jinja2==3.1.3
-tenacity==8.5.0
-pathspec==0.12.1
-pyjwt==2.9.0
-
-# ===== File Processing =====
-binaryornot==0.4.4
-
-# ===== System Utilities =====
-psutil==5.9.6
-
-# ===== HTTP and Network =====
-requests==2.31.0
-urllib3==2.1.0
-
-# ===== EXCLUDED DEPENDENCIES (Cause import errors in HF Spaces) =====
-# docker - Causes "No module named 'docker'" error
-# rapidfuzz - Causes build issues in some environments
-# browsergym-core - Heavy dependency with potential conflicts
-# google-* - All Google dependencies excluded
-# e2b - External service
-# redis - External service
-# tree-sitter - Causes parsing issues
-# openhands-aci - Azure specific, not available
-# whatthepatch - Can cause conflicts
-# protobuf - Version conflicts
-# opentelemetry-* - Heavy monitoring dependencies
-# pygithub - Not essential for basic functionality
-# joblib - Not essential for basic functionality
-# python-socketio - WebSocket not essential for basic API
-# sse-starlette - SSE not essential for basic API
-# fastmcp - MCP not essential for basic functionality
-# python-frontmatter - Not essential for basic functionality
-# PyPDF2, python-pptx, etc. - Document processing not essential
-# boto3 - Cloud storage not essential
-# Pillow - Image processing not essential
-
-# NOTE: This ultra-minimal configuration prioritizes stability
-# and successful deployment over advanced features
