--- requirements_medium.txt
@@ -1,8 +1,0 @@
-# Medium Power Requirements for HF Spaces
-# More features but still lightweight
-
-fastapi==0.104.1
-uvicorn[standard]==0.24.0
-httpx==0.25.2
-pydantic==2.5.0
-python-multipart==0.0.6
