--- requirements_personal.txt
@@ -1,32 +1,0 @@
-# Personal OpenHands Backend - OPENROUTER ONLY! 🎯
-# Perfect for your OpenRouter API key setup
-
-# Core FastAPI
-fastapi==0.115.4
-uvicorn[standard]==0.32.0
-pydantic==2.10.3
-pydantic-settings==2.6.1
-
-# LLM Support - OPENROUTER ONLY!
-litellm==1.52.14
-# NO openai package needed - OpenRouter handles it!
-# NO anthropic package needed - OpenRouter handles it!
-
-# HTTP & Async
-httpx==0.27.2
-aiofiles==24.1.0
-websockets==13.1
-
-# Data Processing (minimal)
-pandas==2.2.3
-numpy==2.1.3
-Pillow==11.0.0
-
-# Utilities
-python-multipart==0.0.12
-python-dotenv==1.0.1
-toml==0.10.2
-packaging==24.2
-typing-extensions==4.12.2
-
-# That's it! Super minimal for OpenRouter setup 🚀
