--- requirements_ultra_simple.txt
@@ -1,7 +1,0 @@
-# Ultra Simple Requirements for HF Spaces
-# Only essential dependencies for FastAPI + OpenRouter
-
-fastapi==0.104.1
-uvicorn[standard]==0.24.0
-httpx==0.25.2
-pydantic==2.5.0
